Move generic FS utils to the right module.
