+++ src/client/common/utils/filesystem.ts
@@ -1,7 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+import * as fs from 'fs';
 import * as path from 'path';
+import * as vscode from 'vscode';
+import { logError } from '../../logging';
 import { getOSType, OSType } from './platform';
 
 /**
@@ -27,4 +30,60 @@ export function areSameFilename(filename1: string, filename2: string): boolean {
     const norm1 = normalizeFilename(filename1);
     const norm2 = normalizeFilename(filename2);
     return norm1 === norm2;
+}
+
+export import FileType = vscode.FileType;
+
+export type DirEntry = {
+    filename: string;
+    filetype: FileType;
+};
+
+interface IKnowsFileType {
+    isFile(): boolean;
+    isDirectory(): boolean;
+    isSymbolicLink(): boolean;
+}
+
+// This helper function determines the file type of the given stats
+// object.  The type follows the convention of node's fs module, where
+// a file has exactly one type.  Symlinks are not resolved.
+export function convertFileType(info: IKnowsFileType): FileType {
+    if (info.isFile()) {
+        return FileType.File;
+    }
+    if (info.isDirectory()) {
+        return FileType.Directory;
+    }
+    if (info.isSymbolicLink()) {
+        // The caller is responsible for combining this ("logical or")
+        // with File or Directory as necessary.
+        return FileType.SymbolicLink;
+    }
+    return FileType.Unknown;
+}
+
+/**
+ * Identify the file type for the given file.
+ */
+export async function getFileType(
+    filename: string,
+    opts: {
+        ignoreErrors: boolean;
+    } = { ignoreErrors: true },
+): Promise<FileType | undefined> {
+    let stat: fs.Stats;
+    try {
+        stat = await fs.promises.lstat(filename);
+    } catch (err) {
+        if (err.code === 'ENOENT') {
+            return undefined;
+        }
+        if (opts.ignoreErrors) {
+            logError(`lstat() failed for "${filename}" (${err})`);
+            return FileType.Unknown;
+        }
+        throw err; // re-throw
+    }
+    return convertFileType(stat);
 }
